fashionclip_experiment/requirements.txt
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
requests>=2.28.0
pymongo>=4.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
